move utils under folder and improve test coverage
